+++ src/main/java/com/backend/locality/api/locality/interfaces/ILocalityRepository.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface ILocalityRepository {
     public List<LocalityModel> findAll();
-    public LocalityModel findLocalityById(int id);
+    public LocalityModel findLocalityById(Long id);
     public LocalityModel saveLocality(LocalityModel locality);
 }
